+++ Models/DoseGuard/Patient.py
@@ -7,7 +7,7 @@ import sqlalchemy as sa
 class Patient(Base):
     __tablename__ = DoseGuardConfig.SQL_PATIENT_TABLE
 
-    id = sa.Column(sa.BigInteger, primary_key=True, autoincrement=True)
+    id = sa.Column(sa.Integer, primary_key=True, autoincrement=True)
     name = sa.Column(sa.String(50), nullable=False)
     contact = sa.Column(sa.String(25))
     age = sa.Column(sa.Integer)
